+++ lib/error.dart
@@ -1,5 +1,5 @@
 import 'builder_helper.dart' as helper;
 import 'parser_builder.dart';
 
-part 'src/error/labeled.dart';
-part 'src/error/malformed.dart';
+part 'src/error/expected.dart';
+part 'src/error/nested.dart';