+++ lib/src/error/nested.dart
@@ -1,19 +1,21 @@
 part of '../../error.dart';
 
-class Malformed<I, O> extends ParserBuilder<I, O> {
-  static const _template = '''
+class Nested<I, O> extends ParserBuilder<I, O> {
+  static const _template = r'''
 {{p1}}
 if (state.ok) {
   {{res}} = {{p1_res}};
 } else if (state.log) {
-  state.error = ErrMalformed(state.pos, const Tag({{tag}}), [state.error]);
+  state.error = ErrNested(state.pos, {{message}}, {{tag}}, [state.error]);
 }''';
+
+  final String message;
 
   final ParserBuilder<I, O> parser;
 
-  final String tag;
+  final String? tag;
 
-  const Malformed(this.tag, this.parser);
+  const Nested(this.message, this.tag, this.parser);
 
   @override
   Map<String, ParserBuilder> getBuilders() {
@@ -24,8 +26,15 @@ if (state.ok) {
 
   @override
   Map<String, String> getTags(Context context) {
+    var tag = 'null';
+    if (this.tag != null) {
+      final value = helper.escapeString(this.tag!, false);
+      tag = 'const Tag(\'$value\')';
+    }
+
     return {
-      'tag': helper.escapeString(tag),
+      'message': helper.escapeString(message),
+      'tag': tag,
     };
   }