+++ lib/src/transformers/char_class_parser.dart
@@ -580,7 +580,7 @@ abstract class Err {
       for (final error in error.errors) {
         _flatten(error, result);
       }
-    } else if (error is ErrWithTagAndErrors) {
+    } else if (error is ErrNested) {
       final inner = <Err>[];
       for (final nested in error.errors) {
         _flatten(nested, inner);
@@ -590,18 +590,13 @@ abstract class Err {
       inner.removeWhere((e) => e.offset < farthest);
       final offset = error.offset;
       final tag = error.tag;
-      result.add(ErrExpected.tag(offset, tag));
+      if (tag != null) {
+        result.add(ErrExpected.tag(offset, tag));
+      }
+
       if (farthest > offset) {
-        if (error is ErrMalformed) {
-          result.add(_ErrInner(farthest, offset, 'Malformed $tag'));
-          result.addAll(inner);
-        } else if (error is ErrNested) {
-          if (farthest > offset) {
-            result.addAll(inner);
-          }
-        } else {
-          throw StateError('Internal error');
-        }
+        result.add(_ErrInner(farthest, offset, error.message));
+        result.addAll(inner);
       }
     } else {
       result.add(error);
@@ -700,34 +695,6 @@ class ErrExpected extends Err {
   }
 }
 
-class ErrMalformed extends ErrWithTagAndErrors {
-  @override
-  final List<Err> errors;
-
-  @override
-  final int offset;
-
-  @override
-  final Tag tag;
-
-  ErrMalformed(this.offset, this.tag, this.errors);
-
-  @override
-  int get length => 1;
-
-  @override
-  // ignore: hash_and_equals
-  bool operator ==(other) {
-    return super == other && other is ErrMalformed;
-  }
-
-  @override
-  String toString() {
-    final result = 'Malformed $tag';
-    return result;
-  }
-}
-
 class ErrMessage extends Err {
   @override
   final int length;
@@ -753,17 +720,18 @@ class ErrMessage extends Err {
   }
 }
 
-class ErrNested extends ErrWithTagAndErrors {
+class ErrNested extends ErrWithErrors {
   @override
   final List<Err> errors;
+
+  final String message;
 
   @override
   final int offset;
 
-  @override
-  final Tag tag;
+  final Tag? tag;
 
-  ErrNested(this.offset, this.tag, this.errors);
+  ErrNested(this.offset, this.message, this.tag, this.errors);
 
   @override
   int get length => 1;
@@ -771,13 +739,15 @@ class ErrNested extends ErrWithTagAndErrors {
   @override
   // ignore: hash_and_equals
   bool operator ==(other) {
-    return super == other && other is ErrNested;
+    return super == other &&
+        other is ErrNested &&
+        other.message == message &&
+        other.tag == tag;
   }
 
   @override
   String toString() {
-    final result = 'Nested $tag';
-    return result;
+    return message;
   }
 }